+++ code/Model/Indexer/Algoliacategories.php
@@ -8,7 +8,6 @@ class Algolia_Algoliasearch_Model_Indexer_Algoliacategories extends Mage_Index_M
     private $engine;
     private $config;
 
-    public static $product_categories = array();
     private static $credential_error = false;
 
     public function __construct()
@@ -177,9 +176,6 @@ class Algolia_Algoliasearch_Model_Indexer_Algoliacategories extends Mage_Index_M
             $this->engine
                 ->rebuildCategoryIndex(null, $updateCategoryIds);
         }
-
-        $process_products = Mage::getSingleton('index/indexer')->getProcessByCode('algolia_search_indexer');
-        $process_products->changeStatus(Mage_Index_Model_Process::STATUS_REQUIRE_REINDEX);
     }
 
     /**