+++ code/Model/Resource/Engine.php
@@ -5,7 +5,8 @@
 class Algolia_Algoliasearch_Model_Resource_Engine extends Mage_CatalogSearch_Model_Resource_Fulltext_Engine
 {
     const ONE_TIME_AMOUNT = 100;
-    private $helper;
+    /** @var Algolia_Algoliasearch_Helper_Logger */
+    private $logger;
     private $queue;
     private $config;
     private $product_helper;
@@ -18,6 +19,7 @@ class Algolia_Algoliasearch_Model_Resource_Engine extends Mage_CatalogSearch_Mod
 
         $this->queue = Mage::getSingleton('algoliasearch/queue');
         $this->config = Mage::helper('algoliasearch/config');
+        $this->logger = Mage::helper('algoliasearch/logger');
         $this->product_helper = Mage::helper('algoliasearch/entity_producthelper');
         $this->category_helper = Mage::helper('algoliasearch/entity_categoryhelper');
         $this->suggestion_helper = Mage::helper('algoliasearch/entity_suggestionhelper');
@@ -29,16 +31,6 @@ class Algolia_Algoliasearch_Model_Resource_Engine extends Mage_CatalogSearch_Mod
             $this->queue->add($observer, $method, $data, $nb_retry);
         else
             Mage::getSingleton($observer)->$method(new Varien_Object($data));
-    }
-
-    public function getAllowedVisibility()
-    {
-        return Mage::getSingleton('catalog/product_visibility')->getVisibleInSearchIds();
-    }
-
-    public function allowAdvancedIndex()
-    {
-        return FALSE;
     }
 
     public function removeProducts($storeId = null, $product_ids = null)
@@ -95,6 +87,12 @@ class Algolia_Algoliasearch_Model_Resource_Engine extends Mage_CatalogSearch_Mod
     {
         foreach (Mage::app()->getStores() as $store)
         {
+            if ($this->config->isEnabledBackEnd($store->getId()) === false)
+            {
+                $this->logger->log('INDEXING IS DISABLED FOR '. $this->logger->getStoreName($store->getId()));
+                continue;
+            }
+
             $this->addToQueue('algoliasearch/observer', 'rebuildPageIndex', array('store_id' => $store->getId()), $this->config->getQueueMaxRetries());
         }
     }
@@ -103,6 +101,12 @@ class Algolia_Algoliasearch_Model_Resource_Engine extends Mage_CatalogSearch_Mod
     {
         foreach (Mage::app()->getStores() as $store)
         {
+            if ($this->config->isEnabledBackEnd($store->getId()) === false)
+            {
+                $this->logger->log('INDEXING IS DISABLED FOR '. $this->logger->getStoreName($store->getId()));
+                continue;
+            }
+
             $this->addToQueue('algoliasearch/observer', 'rebuildAdditionalSectionsIndex', array('store_id' => $store->getId()), $this->config->getQueueMaxRetries());
         }
     }
@@ -111,6 +115,12 @@ class Algolia_Algoliasearch_Model_Resource_Engine extends Mage_CatalogSearch_Mod
     {
         foreach (Mage::app()->getStores() as $store)
         {
+            if ($this->config->isEnabledBackEnd($store->getId()) === false)
+            {
+                $this->logger->log('INDEXING IS DISABLED FOR '. $this->logger->getStoreName($store->getId()));
+                continue;
+            }
+
             $size       = $this->suggestion_helper->getSuggestionCollectionQuery($store->getId())->getSize();
             $by_page    = $this->config->getNumberOfElementByPage();
             $nb_page    = ceil($size / $by_page);
@@ -130,10 +140,14 @@ class Algolia_Algoliasearch_Model_Resource_Engine extends Mage_CatalogSearch_Mod
 
     public function rebuildProducts()
     {
-        Mage::getSingleton('algoliasearch/observer')->saveSettings();
-
         foreach (Mage::app()->getStores() as $store)
         {
+            if ($this->config->isEnabledBackEnd($store->getId()) === false)
+            {
+                $this->logger->log('INDEXING IS DISABLED FOR '. $this->logger->getStoreName($store->getId()));
+                continue;
+            }
+
             if ($store->getIsActive())
             {
                 $this->_rebuildProductIndex($store->getId(), array());
@@ -147,10 +161,14 @@ class Algolia_Algoliasearch_Model_Resource_Engine extends Mage_CatalogSearch_Mod
 
     public function rebuildCategories()
     {
-        Mage::getSingleton('algoliasearch/observer')->saveSettings();
-
         foreach (Mage::app()->getStores() as $store)
         {
+            if ($this->config->isEnabledBackEnd($store->getId()) === false)
+            {
+                $this->logger->log('INDEXING IS DISABLED FOR '. $this->logger->getStoreName($store->getId()));
+                continue;
+            }
+
             if ($store->getIsActive())
             {
                 $this->addToQueue('algoliasearch/observer', 'rebuildCategoryIndex', array('store_id' => $store->getId(), 'category_ids' =>  array()), $this->config->getQueueMaxRetries());
@@ -220,11 +238,15 @@ class Algolia_Algoliasearch_Model_Resource_Engine extends Mage_CatalogSearch_Mod
         else
             $this->addToQueue('algoliasearch/observer', 'rebuildProductIndex', array('store_id' => $storeId, 'product_ids' =>  $productIds), $this->config->getQueueMaxRetries());
 
+
         return $this;
     }
 
     public function prepareEntityIndex($index, $separator = ' ')
     {
+        if ($this->config->isEnabledBackEnd(Mage::app()->getStore()->getId()) === false)
+            return parent::rebuildIndex($index, $separator);
+
         foreach ($index as $key => $value) {
             if (is_array($value) && ! empty($value)) {
                 $index[$key] = join($separator, array_unique(array_filter($value)));
@@ -238,10 +260,5 @@ class Algolia_Algoliasearch_Model_Resource_Engine extends Mage_CatalogSearch_Mod
     public function saveSettings()
     {
         Mage::getSingleton('algoliasearch/observer')->saveSettings();
-    }
-
-    public function test()
-    {
-        return true;
     }
 }