+++ dev/bin/start.sh
@@ -13,7 +13,6 @@ n98-magerun --root-dir=/var/www/htdocs config:set algoliasearch/credentials/api_
 n98-magerun --root-dir=/var/www/htdocs config:set algoliasearch/credentials/index_prefix $INDEX_PREFIX
 n98-magerun --root-dir=/var/www/htdocs config:set web/unsecure/base_url $BASE_URL
 n98-magerun --root-dir=/var/www/htdocs config:set web/secure/base_url $BASE_URL
-n98-magerun --root-dir=/var/www/htdocs config:set dev/log/active $BASE_URL
 
 # GET / to initialize the algolia_search_indexer (Open to another cleaner way to do it :) )
 wget $BASE_URL
@@ -23,6 +22,9 @@ n98-magerun --root-dir=/var/www/htdocs index:reindex algolia_search_indexer
 n98-magerun --root-dir=/var/www/htdocs index:reindex algolia_search_indexer_cat
 n98-magerun --root-dir=/var/www/htdocs index:reindex algolia_search_indexer_pages
 n98-magerun --root-dir=/var/www/htdocs index:reindex search_indexer_suggest
+
+# do it after indexing so that var/log doesn't get created as root
+n98-magerun --root-dir=/var/www/htdocs config:set dev/log/active 1
 
 service apache2 stop
 exec /usr/sbin/apache2ctl -D FOREGROUND